requirements.txt
keras==2.2.0
tensorflow==1.4.0
numpy==1.14.5
pillow==5.2.0
matplotlib==2.2.2
opencv-python==3.4.1.15
configparser==3.5.0
graphviz==0.8.4
pydot==1.2.4
pytorch-cpu==1.0.0
